Create helper functions file within rpi
